Refactor schedule screen into modular components
