+++ lib/pages/schedule_screen.dart
@@ -6,7 +6,8 @@ import '../models/event.dart' show Event;
 import '../services/data_provider_service.dart';
 import '../widgets/active_tab_indicator.dart';
 import '../widgets/momentum_appbar.dart';
-import '../widgets/schedule_item.dart';
+import '../widgets/schedule/day_tab.dart';
+import '../widgets/schedule/event_day_list.dart';
 
 class ScheduleScreen extends StatefulWidget {
   const ScheduleScreen({super.key});
@@ -91,61 +92,21 @@ class ScheduleScreenState extends State<ScheduleScreen>
           ? const Center(child: CircularProgressIndicator())
           : TabBarView(
               controller: _tabController,
-              children:
-                  uniqueDays.map((day) => getEventDay(day, context)).toList(),
+              children: uniqueDays
+                  .map((day) => EventDayList(
+                        day: day,
+                        allEvents: schedule,
+                        dataProviderService: _dataProviderService,
+                        onRefresh: (value) {
+                          setState(() {
+                            updateTabController(value);
+                            schedule = value;
+                          });
+                        },
+                      ))
+                  .toList(),
             ),
     );
-  }
-
-  // A function that filters events for a given day, and returns a RefreshIndicator
-  // containing a ListView of ScheduleItem widgets based on the filtered events.
-  Widget getEventDay(DateTime day, BuildContext context) {
-    final events = filterEventsForDay(day);
-    return RefreshIndicator(
-      onRefresh: () {
-        return _dataProviderService
-            .getSchedule(forceNewData: true)
-            .then((value) {
-          setState(() {
-            updateTabController(value);
-            schedule = value;
-          });
-        });
-      },
-      child: ListView.builder(
-        clipBehavior: Clip.antiAlias,
-        padding: const EdgeInsets.all(0),
-        itemCount: events.length,
-        itemBuilder: (context, index) {
-          final event = events[index];
-          final color = event.category != null
-              ? Color(event.category?.color.hex ?? 0xFFFFFF)
-              : Theme.of(context).primaryColor;
-          final textColor = event.category != null
-              ? Color(event.category?.textColor.hex ?? 0xFFFFFF)
-              : Theme.of(context).colorScheme.onPrimary;
-          return ScheduleItem(
-            event: event,
-            color: color,
-            textColor: textColor,
-          );
-        },
-      ),
-    );
-  }
-
-  // Filter events in the schedule for a specific day.
-  // Adjust the event date for comparison if it's before 3 AM.
-  List<Event> filterEventsForDay(DateTime day) {
-    return schedule.where((event) {
-      DateTime eventDate = event.start;
-      // Adjust the event date for comparison
-      if (eventDate.hour < 3) {
-        eventDate = eventDate.subtract(const Duration(days: 1));
-      }
-      eventDate = DateTime(eventDate.year, eventDate.month, eventDate.day);
-      return eventDate == day;
-    }).toList();
   }
 
   PreferredSizeWidget getTabBar(BuildContext context) {
@@ -158,43 +119,16 @@ class ScheduleScreenState extends State<ScheduleScreen>
       indicatorPadding: const EdgeInsets.all(0),
       dividerHeight: 0,
       tabs: uniqueDays.map((day) {
-        final weekday =
-            DateFormat('EEE').format(day).replaceAll(RegExp(r'\.$'), '');
-        final dayOfMonth = DateFormat('d').format(day);
-
         return Tab(
           child: AnimatedBuilder(
             animation: _tabController.animation!,
             builder: (context, child) {
               final color = Theme.of(context).colorScheme.onPrimary;
-
-              return SizedBox(
-                width: deviceWidth / uniqueDays.length,
-                height: 50,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Text(
-                      ((deviceWidth > 365 || weekday.length == 2)
-                              ? weekday
-                              : weekday[0] + weekday[2])
-                          .toUpperCase(),
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        color: color,
-                        fontSize: 12,
-                      ),
-                    ),
-                    Text(
-                      dayOfMonth,
-                      style: TextStyle(
-                        fontSize: 16,
-                        color: color,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ],
-                ),
+              return DayTab(
+                day: day,
+                color: color,
+                deviceWidth: deviceWidth,
+                daysCount: uniqueDays.length,
               );
             },
           ),
